Allow table of contents field to default on any content entity type
The custom field type needs to be able to render for any content entity
type in the default case (i.e. no nid set). Note that it will only
target the fields of the parent entity, not a super-parent. The
non-default case (i.e. with nid set) will still only target Node
entities.